Move pkg/catalog to pkg/output/catalog
