+++ pkg/legacy/create-component-entity.go
@@ -7,9 +7,9 @@ import (
 	"time"
 
 	bscatalog "github.com/giantswarm/backstage-catalog-importer/pkg/bscatalog/v1alpha1"
-	"github.com/giantswarm/backstage-catalog-importer/pkg/catalog/component"
 	"github.com/giantswarm/backstage-catalog-importer/pkg/input/helmchart"
 	"github.com/giantswarm/backstage-catalog-importer/pkg/input/repositories"
+	"github.com/giantswarm/backstage-catalog-importer/pkg/output/catalog/component"
 )
 
 // CreateComponentEntity is the deprecated way of generating a component entity.
