+++ pkg/output/catalog/component/component.go
