+++ pkg/output/catalog/component/component_options.go
