+++ pkg/output/catalog/component/component_test.go
