+++ pkg/output/catalog/component/to_entity.go
